Exclude packages by package name instead of source_name
Currently when using repos, both include and exclude work
based on source_name. However, i think the most common case
is to include a number of packages based on source name and
exclude via subpackage names to slim dependencies.

This patch implements this behavior. An alternative would be
to have several sections to exclude and include sources and
subpackage but i think we will implement it later if it's really
required.

Change-Id: Ibfc47c8f6dc00b4d82fa99a690b231c92167938d