Include/Exclude packages by package name instead of source_name
Currently when using repos, both include and exclude work
based on source_name. However, in order to create light repository,
it's easier to include or exclude based on subpackage names.

This patch implements this behavior. An alternative would be
to have several sections to exclude and include sources and
subpackage but i think we will implement it later if it's really
required.

Change-Id: I2a13e7e9639676cae01cdaeb6e11553275dabbf1