Add support for exclude subpackages from koji downloader
In order to create repos as small as possible using sinkhole we
need to be able to exclude subpackages from a specific koji build.

This patch adds that capacity by adding a new exclude field to the
koji constraints file. To implement it, I'm using the same filter
methods used for repository synchronization. However, as packages
in koji plugin are represented as a dict i need to move the result
to lists instead of sets (dicts are not suported in sets).

Change-Id: Id7206192be79d509587a7539d3a810c9ef3a488b